fetch data for location
